dyesub:  Convert now-defunct use of the 'PT1' macro to normal 'PT'
Before we could have non-integer point dimensions, we used the PT1
macro to round up a given dimension to the next highest integer value.
This matters because the dyesub models are nearly all full-bleed, and the
truncation/round-down could result in white margins.

When we added fractional dimensions, the PT1 macro became superfluous, and
was #defined to be the same as the original PT macro.

After that was converted, create a new PT1 macro that is defined as
PT(x) + 1 The intent is to use this to differentiate print sizes that
have the same on-printer dimensions but act differently, eg having additional
cuts.